Links, identifiers and local objects and properties
Introduced local objects and properties. Local objects and propertes
don't have to be specified in global vocabulary.

Also added identifiers in order to know how to identify same object in
order to be able to update same object.

And also introduced links, that should point to a URI from a know
vocabulary.