Refactor RC AR data structures
RC AR data are very important and are referenced from multiple other
datasets, so AR must be very clean and eassy to use.

The main issue with AR was, that data was split into graphical and
textual. Now both graphical and textual are represented as on model.

Another issue was, that this data was split into many multiple datasets,
now all AR models belong to a single dataset.